lvl3/requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
psycopg2
Pillow>=5.3.0,<5.4.0

django-rest-auth
django-allauth
